requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# Machine Learning Libraries
xgboost==1.7.6
lightgbm==4.0.0
tensorflow==2.13.0
torch==2.0.1
transformers==4.32.1

# Technical Analysis
TA-Lib==0.4.27
ta==0.10.2

# Data Collection APIs
yfinance==0.2.18
alpha-vantage==2.3.1
python-binance==1.0.19
MetaTrader5==5.0.45

# Backtesting
backtrader==1.9.78.123
zipline-reloaded==3.0.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Hyperparameter Optimization
optuna==3.2.0
hyperopt==0.2.7

# Model Management
mlflow==2.5.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Utilities
tqdm==4.65.0
joblib==1.3.1
requests==2.31.0

# Risk Management
pyfolio==0.9.2
empyrical==0.5.5

# Time Series
statsmodels==0.14.0

# Advanced Techniques
shap==0.42.1
umap-learn==0.5.3

# Ensemble Methods and Advanced ML
scipy==1.11.1


# Monitoring and Alerting
schedule==1.2.0
